Fixed URL generation/redirect issues
- I've altered the App Overview page so it functions as a main admin dashboard
- Adjusted URL rules so authentication controllers redirect to the dashboard rather than `/undefined`
- Deleted the `PageController` in the app folder as this was causing conflicts, the page methods are now in their own controllers
- The above changed fixed the URL generation issues for the nav bar
