+++ Controllers/Authentication/ResetPasswordController.cs
@@ -36,8 +36,8 @@ namespace Deepcove_Trust_Website.Controllers.Authentication
             {
                 return Redirect(Url.Action(
                     "Index",
-                    "Page",
-                    new { area = "admin-portal,web" }
+                    "Dashboard",
+                    new { area = "admin" }
                 ));
             }
