+++ Resources/js/React/AdminPortal/Dashboard.js
@@ -1,6 +1,6 @@
 ﻿import React, { Component, Fragment } from 'react';
 import { render } from 'react-dom';
-import { StaticAlert } from '../../Components/Alert';
+import { StaticAlert } from '../Components/Alert';
 
 export default class Overview extends Component {
     render() {
@@ -8,18 +8,19 @@ export default class Overview extends Component {
             <Fragment>
                 <h1>Reactified Overview</h1>
                 <StaticAlert type="danger">Coming Soon</StaticAlert>
-                <strong>This page will show</strong>
+                <strong>Admin dashboard will show</strong>
                 <ul>
                     <li>Number of Quizes</li>
                     <li>Number of Fact Files and categories</li>
                     <li>Number of Walks and their activities</li>
                     <li>Overview of urgent notices</li>
+                    <li>Some web stats</li>
                 </ul>
             </Fragment>
         );
     }
 }
 
-if (document.getElementById('react_app_overview')) {
-    render(<Overview />, document.getElementById('react_app_overview'));
+if (document.getElementById('react_admin_dashboard')) {
+    render(<Overview />, document.getElementById('react_admin_dashboard'));
 }
