--- Views/AdminPortal/App/Overview.cshtml
@@ -1,8 +1,0 @@
-﻿
-@{
-    ViewData["Title"] = "Overview";
-    Layout = "~/Views/Shared/_Layout.cshtml";
-}
-
-    <div id="react_app_overview"></div>
-
