+++ Views/AdminPortal/Dashboard.cshtml
@@ -1,0 +1,8 @@
+﻿
+@{
+    ViewData["Title"] = "Admin Dashboard";
+    Layout = "~/Views/Shared/_Layout.cshtml";
+}
+
+    <div id="react_admin_dashboard"></div>
+
