+++ Views/Shared/_AdminNav.cshtml
@@ -24,8 +24,13 @@
                 <li class="nav-item dropdown">
                     <a data-toggle="dropdown" aria-expanded="false" class="dropdown-toggle nav-link"> Discover Deep Cove</a>
                     <div role="menu" class="dropdown-menu">
-                        <a class="dropdown-item" role="presentation" asp-area="admin" asp-controller="Page" asp-action="FactFiles">Fact Files</a>
-                        <a class="dropdown-item" role="presentation" asp-area="admin" asp-controller="Page" asp-action="Quizzes">Quizes</a>
+                        @*<a class="dropdown-item" role="presentation" asp-area="admin" asp-controller="AppSettings">Settings</a>*@
+                        <div class="dropdown-divider"></div>
+                        <h6 class="dropdown-header">Manage Content</h6>
+                        <a class="dropdown-item" role="presentation" asp-area="admin" asp-controller="ManageFactFiles">Fact Files</a>
+                        @*<a class="dropdown-item" role="presentation" asp-area="admin" asp-controller="ManageQuizzes">Quizes</a>
+                        <a class="dropdown-item" role="presentation" asp-area="admin" asp-controller="ManageTracks">Tracks</a>*@
+                        
                     </div>
                 </li>
